Move common song select implementation for online play
